+++ eras/conway/impl/src/Cardano/Ledger/Conway/Governance/Procedures.hs
@@ -55,6 +55,7 @@ module Cardano.Ledger.Conway.Governance.Procedures (
   indexedGovProps,
   Constitution (..),
   constitutionAnchorL,
+  constitutionGuardrailsScriptHashL,
   constitutionScriptL,
   showGovActionType,
   -- Lenses
@@ -798,7 +799,7 @@ data GovAction era
       !(StrictMaybe (GovPurposeId 'PParamUpdatePurpose))
       -- | Proposed changes to PParams
       !(PParamsUpdate era)
-      -- | Policy hash protection
+      -- | Guardrails script hash protection
       !(StrictMaybe ScriptHash)
   | HardForkInitiation
       -- | Previous governance action id of `HardForkInitiation` type, which corresponds
@@ -809,7 +810,7 @@ data GovAction era
   | TreasuryWithdrawals
       -- | Proposed treasury withdrawals
       !(Map RewardAccount Coin)
-      -- | Policy hash protection
+      -- | Guardrails script hash protection
       !(StrictMaybe ScriptHash)
   | NoConfidence
       -- | Previous governance action id of `NoConfidence` or `UpdateCommittee` type, which
@@ -892,7 +893,7 @@ instance EraPParams era => EncCBOR (GovAction era) where
 
 data Constitution era = Constitution
   { constitutionAnchor :: !Anchor
-  , constitutionScript :: !(StrictMaybe ScriptHash)
+  , constitutionGuardrailsScriptHash :: !(StrictMaybe ScriptHash)
   }
   deriving (Generic, Ord)
   deriving (ToJSON) via KeyValuePairs (Constitution era)
@@ -900,8 +901,12 @@ data Constitution era = Constitution
 constitutionAnchorL :: Lens' (Constitution era) Anchor
 constitutionAnchorL = lens constitutionAnchor (\x y -> x {constitutionAnchor = y})
 
+constitutionGuardrailsScriptHashL :: Lens' (Constitution era) (StrictMaybe ScriptHash)
+constitutionGuardrailsScriptHashL = lens constitutionGuardrailsScriptHash (\x y -> x {constitutionGuardrailsScriptHash = y})
+
+{-# DEPRECATED constitutionScriptL "In favor of `constitutionGuardrailsScriptHashL`" #-}
 constitutionScriptL :: Lens' (Constitution era) (StrictMaybe ScriptHash)
-constitutionScriptL = lens constitutionScript (\x y -> x {constitutionScript = y})
+constitutionScriptL = constitutionGuardrailsScriptHashL
 
 instance Era era => FromJSON (Constitution era) where
   parseJSON = withObject "Constitution" $ \o ->
@@ -913,7 +918,7 @@ instance ToKeyValuePairs (Constitution era) where
   toKeyValuePairs c@(Constitution _ _) =
     let Constitution {..} = c
      in ["anchor" .= constitutionAnchor]
-          <> ["script" .= cScript | SJust cScript <- [constitutionScript]]
+          <> ["script" .= cScript | SJust cScript <- [constitutionGuardrailsScriptHash]]
 
 deriving instance Eq (Constitution era)
 
@@ -934,7 +939,7 @@ instance Era era => EncCBOR (Constitution era) where
     encode $
       Rec (Constitution @era)
         !> To constitutionAnchor
-        !> E (encodeNullStrictMaybe encCBOR) constitutionScript
+        !> E (encodeNullStrictMaybe encCBOR) constitutionGuardrailsScriptHash
 
 instance Era era => ToCBOR (Constitution era) where
   toCBOR = toEraCBOR @era